Restructuring commit
This changes the object to having just a single constructor that includes the
CSV filename, and a boolean about whether there's a header or not. We could try
to be really clever and make an intelligent guess, but that guess could be
wrong.

This also gets rid of the fields/field_num mess .. if the file had a header,
you can use name or column number, and if not, just column number will work.